+++ libs/cardano-ledger-core/src/Cardano/Ledger/UMapCompact.hs
@@ -1,6 +1,0 @@
-module Cardano.Ledger.UMapCompact
-  {-# DEPRECATED "Use Cardano.Ledger.UMap instead" #-}
-  (module X)
-where
-
-import Cardano.Ledger.UMap as X